Move PDFReaderState.AnnotationKey to PDFReaderAnnotationKey
